chore: fix test failures and staticcheck warnings (#154)
- internal/web/web.go: restore else-chain so SPA route fallback doesn't
  hit the generic 404 — IsSPARoute branch now correctly serves index.html
- internal/importers/hevy/transform.go: remove unused intPtr/floatPtr
- graphify-out: refresh knowledge graph

+++ internal/foodimport/foodimport.go
@@ -266,7 +266,7 @@ func localFingerprint(path string, filter ports.BulkFilter) (string, error) {
 		Size       int64
 		ModifiedNS int64
 		Filter     ports.BulkFilter
-	}{abs, uint64(stat.Dev), stat.Ino, info.Size(), info.ModTime().UnixNano(), filter})
+	}{abs, stat.Dev, stat.Ino, info.Size(), info.ModTime().UnixNano(), filter})
 	if err != nil {
 		return "", fmt.Errorf("encode fingerprint: %w", err)
 	}
